requirements.txt
streamlit
openai
scikit-learn
genai
google.generativeai